pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "matrix-proxy-bot"
version = "0.1.0"
description = "Matrix ↔ Agent-Shell webhook relay bot"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Edd Wilder-James", email = "edd@example.com"}
]
requires-python = ">=3.13"
dependencies = [
    "matrix-nio[e2e]>=0.26.1",
    "aiosqlite>=0.20.0",
    "fastapi>=0.114.0",
    "uvicorn>=0.27.0",
    "httpx>=0.26.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/matrix_proxy_bot"]
